requirements.txt
anthropic>=0.40.0
mem0ai>=0.1.0
chromadb>=0.5.0
elevenlabs>=1.0.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
httpx>=0.27.0
sentence-transformers>=3.0.0
groq>=1.1.2
psutil>=5.9.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
croniter>=1.4.0